% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yenting.R
\name{game}
\alias{game}
\title{Get the card repeatedly}
\usage{
game(card, point)
}
\arguments{
\item{card}{The first vector.}

\item{point}{The max total which you want stop to pick new one card}
}
\value{
All of your card and whether you win the game or not
}
\description{
You will get several card randomly until the total sum is larger than you setting in the beginning.
}
