% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yenting.R
\name{PermuTest}
\alias{PermuTest}
\title{Permutation test}
\usage{
PermuTest(data_A, data_B, n, alpha)
}
\arguments{
\item{n}{The times of simulation.}

\item{alpha}{Significant level.}

\item{dataA}{The first vector.}

\item{dataB}{The second vector.}
}
\value{
reject or not reject
}
\description{
Take any two vector you want to do the permutation test
}
